+++ lib/simple_sizer.dart
@@ -2,9 +2,10 @@ library simple_sizer;
 
 //From Folder exports
 export 'src/exports/size_config.dart';
-export 'src/exports/reactive_row_col.dart';
-export 'src/exports/reactive_home.dart';
-export 'src/exports/shape_helper.dart';
+
+//From Folder helper_methods
+export 'src/exports/helper_methods/shape_helper.dart';
+export 'src/exports/helper_methods/clamp_width_height_helper.dart';
 
 //From Folder extensions
 export 'src/exports/extensions/build_ctx.dart';
@@ -20,4 +21,10 @@ export 'src/exports/extensions/stba_ext.dart';
 export 'src/exports/extensions/text_ext.dart';
 
 //From Folder reactive_widgets
-export 'src/reactive_widgets/reactive_destination.dart';
+export 'src/exports/reactive_widgets/reactive_destination.dart';
+export 'src/exports/reactive_widgets/reactive_row_col.dart';
+export 'src/exports/reactive_widgets/reactive_home.dart';
+
+//From Folder enums
+export 'src/exports/enums/text_align_enum.dart';
+export 'src/exports/enums/align_cntr_enum.dart';