+++ lib/src/exports/extensions/sizedbox_ext.dart
@@ -11,3 +11,14 @@ extension SizedBoxExt on num {
   /// `wbox` extension to create a `SizedBox` widget with that width.
   SizedBox get wbox => SizedBox(width: toDouble());
 }
+
+extension SizedBoxExtWidget on Widget {
+  /// The `szdbox` function returns a `SizedBox` widget with the specified height, width, and child.
+  ///
+  /// Args:
+  ///   h (double): The height of the SizedBox.
+  ///   w (double): The parameter "w" in the above code represents the desired width of the SizedBox
+  /// widget.
+  SizedBox szdbox({double? h, double? w}) =>
+      SizedBox(height: h, width: w, child: this);
+}