+++ lib/src/private/views/reactive_bigscreen_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:simple_sizer/src/reactive_widgets/reactive_destination.dart';
-import 'package:simple_sizer/src/reactive_widgets/reactive_helper.dart';
+import 'package:simple_sizer/src/exports/reactive_widgets/reactive_destination.dart';
+import 'package:simple_sizer/src/private/reactive_helper.dart';
 
 class ReactiveBigScreenView extends StatefulWidget {
   const ReactiveBigScreenView({